+++ backend/ml_logic/parameters.py
@@ -1,0 +1,15 @@
+SEED = 137
+#  the range from segmentation
+# [40:195,35:225,26:137]
+# constants to crop images
+
+MIN_WIDTH = 25 # 40
+MAX_WIDTH = 215 # 195
+
+MIN_HEIGHT = 35 # 35
+MAX_HEIGHT = 225 # 225
+
+MIN_DEPTH = 15 # 26
+MAX_DEPTH = 143 # 137
+
+cache_path ='/home/yaoyx001/code/cache' #cache of nii files
