--- ml_logic/result.py
@@ -1,34 +1,0 @@
-import matplotlib.pyplot as plt
-import pandas as pd
-from sklearn.metrics import classification_report
-
-def plot_loss(history):
-    fig, (ax1, ax2) = plt.subplots(1,2, figsize=(13,4))
-    ax1.plot(history.history['loss'])
-    ax1.plot(history.history['val_loss'])
-    ax1.set_title('Model loss')
-    ax1.set_ylabel('Loss')
-    ax1.set_xlabel('Epoch')
-    ax1.set_ylim(ymin=0, ymax=1)
-    ax1.legend(['Train', 'Validation'], loc='best')
-    ax1.grid(axis="x",linewidth=0.2)
-    ax1.grid(axis="y",linewidth=0.2)
-
-    ax2.plot(history.history['accuracy'])
-    ax2.plot(history.history['val_accuracy'])
-    ax2.set_title('Accuracy')
-    ax2.set_ylabel('Accuracy')
-    ax2.set_xlabel('Epoch')
-    ax2.set_ylim(ymin=0, ymax=1)
-    ax2.legend(['Train', 'Validation'], loc='best')
-    ax2.grid(axis="x",linewidth=0.2)
-    ax2.grid(axis="y",linewidth=0.2)
-
-    plt.show()
-
-def plot_confusion_matrix(model, X_test,y_test):
-    predictions = (model.predict(X_test) > 0.5).astype("int32")
-    predictions = pd.DataFrame(predictions)
-    target_names = ['class 0', 'class 1']
-    print(classification_report(y_test, predictions, target_names=target_names))
-
